Several code improvements; UI improvements; added file cleaner and credit panels.

+++ src/main/java/net/dbd/demode/util/ByteUtil.java
@@ -33,7 +33,7 @@ public class ByteUtil {
     }
 
     public static void toLittleEndian(long num, byte[] result) {
-        toByteArray(num, Integer.BYTES, true, result);
+        toByteArray(num, Long.BYTES, true, result);
     }
 
     private static byte[] toByteArray(long byteSeq, int numBytes, boolean toLittleEndian) {
